+++ tests/testthat/test-paste_linter.R
@@ -8,7 +8,8 @@ test_that("paste_linter skips allowed usages for sep=''", {
   expect_no_lint("sep <- ''; paste('a', sep)", linter)
   expect_no_lint("paste(sep = ',', '', 'a')", linter)
   expect_no_lint("paste0('a', 'b', 'c')", linter)
-  expect_no_lint("expression(paste('a', b, sep = ''))", linter)
+  expect_no_lint("expression(2); paste('a', 'b', 'c', sep = ',')", linter)
+  expect_no_lint("paste('a', 'b', expression(2), sep = ',')", linter)
 })
 
 test_that("paste_linter blocks simple disallowed usages for sep=''", {
@@ -31,8 +32,13 @@ test_that("paste_linter blocks simple disallowed usages for sep=''", {
   )
 
   expect_lint(
-    "c(expression(2), paste('a', 'b', sep = ''))",
-    rex::rex('paste0(...) is better than paste(..., sep = "").'),
+    "c(expression(paste('a', 'b', sep = ',')))",
+    rex::rex("inside expression(...), paste does not accept a 'sep' argument."),
+    paste_linter()
+  )
+  expect_lint(
+    "c(expression(paste('a', 'b', sep = '')))",
+    rex::rex("inside expression(...), paste does not accept a 'sep' argument."),
     paste_linter()
   )
 })